revert PR 11907 which introduced missing FAB bug
