+++ src/pages/home/sidebar/SidebarScreen/index.js
@@ -1,41 +1,52 @@
 import React from 'react';
-import sidebarPropTypes from './sidebarPropTypes';
+import {withOnyx} from 'react-native-onyx';
+import compose from '../../../../libs/compose';
+import withWindowDimensions from '../../../../components/withWindowDimensions';
+import withLocalize from '../../../../components/withLocalize';
+import ONYXKEYS from '../../../../ONYXKEYS';
+import {sidebarPropTypes, sidebarDefaultProps} from './sidebarPropTypes';
 import BaseSidebarScreen from './BaseSidebarScreen';
-import PopoverModal from './PopoverModal';
 
 const SidebarScreen = (props) => {
-    let popoverModal = null;
+    let baseSidebarScreen = null;
 
     /**
      * Method create event listener
      */
     const createDragoverListener = () => {
-        document.addEventListener('dragover', popoverModal.hideCreateMenu);
+        document.addEventListener('dragover', baseSidebarScreen.hideCreateMenu);
     };
 
     /**
      * Method remove event listener.
      */
     const removeDragoverListener = () => {
-        document.removeEventListener('dragover', popoverModal.hideCreateMenu);
+        document.removeEventListener('dragover', baseSidebarScreen.hideCreateMenu);
     };
-
     return (
-        <>
-            <BaseSidebarScreen
-                // eslint-disable-next-line react/jsx-props-no-spreading
-                {...props}
-            />
-            <PopoverModal
-                ref={el => popoverModal = el}
-                onShowCreateMenu={createDragoverListener}
-                onHideCreateMenu={removeDragoverListener}
-            />
-        </>
+        <BaseSidebarScreen
+            ref={el => baseSidebarScreen = el}
+            onShowCreateMenu={createDragoverListener}
+            onHideCreateMenu={removeDragoverListener}
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            {...props}
+        />
     );
 };
 
 SidebarScreen.propTypes = sidebarPropTypes;
+SidebarScreen.defaultProps = sidebarDefaultProps;
 SidebarScreen.displayName = 'SidebarScreen';
 
-export default SidebarScreen;
+export default compose(
+    withLocalize,
+    withWindowDimensions,
+    withOnyx({
+        allPolicies: {
+            key: ONYXKEYS.COLLECTION.POLICY,
+        },
+        betas: {
+            key: ONYXKEYS.BETAS,
+        },
+    }),
+)(SidebarScreen);
