+++ src/pages/home/sidebar/SidebarScreen/sidebarPropTypes.js
@@ -1,8 +1,26 @@
 import PropTypes from 'prop-types';
+import {windowDimensionsPropTypes} from '../../../../components/withWindowDimensions';
+import {withLocalizePropTypes} from '../../../../components/withLocalize';
 
 const sidebarPropTypes = {
 
-    /** reportID in the current navigation state */
-    reportIDFromRoute: PropTypes.string,
+    /** The list of policies the user has access to. */
+    allPolicies: PropTypes.shape({
+        /** The policy name */
+        name: PropTypes.string,
+    }),
+
+    /* Beta features list */
+    betas: PropTypes.arrayOf(PropTypes.string),
+
+    ...windowDimensionsPropTypes,
+
+    ...withLocalizePropTypes,
 };
-export default sidebarPropTypes;
+
+const sidebarDefaultProps = {
+    allPolicies: {},
+    betas: [],
+};
+
+export {sidebarPropTypes, sidebarDefaultProps};
